+++ services/tileserver/CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## [Unreleased]
 
+- `?classes=Alunite,Muscovite` shorthand on categorical raster layers
 - Carto v2 map tiles support
 - Map ingestion tiles fixes
 - Map bounds tile layer